Fixed bug that allowed file mismatch in cluster.split
Count and Name files were allowed to include names not in fasta file. This occurred if user failed to remove chimeras or contaminants by leaving the name or count file off of the remove.seqs or remove.lineage commands.  Not removing the chimeras from the count or name file allows them to be added to the list file because mothur creates the "unique" list from those names.  This fix is designed to make it harder for people to make clustering mistakes that could seriously effect their results because it 3am and they forgot to include a file on remove.seqs and ignored a few the error messages along the way.  We build the "unique" list based on the count or names file to accommodate the sparse distance matrix where sequences may be present in the names or count file, but absent from the distance matrix because their distances are above the cutoff.

#190