Fixes cluster.split bug with sequences not assigned to group
When I added the changes for the vsearch methods I added checks to make sure there were no singletons in that splitting case.  The sequences mothur is complaining about have a tax level below the cutoff, and so they are not assigned to a group and should be singletons.   I remembered the below the distance cutoff  but forgot about the below the taxlevel cutoff.