Add tree algorithm
This algorithm merges sequences together if they are one off of a more
abundant sequence and a specified threshold different in abundance from
a more abundant sequence. In contrast to the simple and unoise
algorithms, the tree algorithm will merge sequences that are more than
one off of the parent sequence if it is a rarer variant of a one off,
two off, etc. The algorithm assumes that mutations build off of each
other and are not independent of each other. The abundance threshold
is set by the delta parameter.