Add deblur algorithm
This algorithm removes counts from sequences that are suspected of being
sequencing variants of other sequences in the dataset. This
implementation generally gives the same result as that of original
python implementation, but does not remove rare sequences, align
sequences, require sequence trimming, or do chimera screening. We also
allow users to define the error distribution using a binomial
distribution. It's also wicked fast.